requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.7
exceptiongroup==1.3.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pluggy==1.6.0
Pygments==2.20.0
pytest==9.0.3
pytest-html==4.2.0
pytest-metadata==3.1.1
python-dotenv==1.2.2
requests==2.33.1
tomli==2.4.1
typing_extensions==4.15.0
urllib3==2.6.3
